+++ src/error-message.ts
@@ -73,7 +73,7 @@ export class ErrorMessage {
   public static ComplexNumberExpected = 'Complex number expected.'
   public static ShouldBeIorJ = 'Should be \'i\' or \'j\'.'
   public static SizeMismatch = 'Array dimensions mismatched.'
-  public static TextJoinResultTooLong = 'TEXTJOIN result exceeds the maximum allowed length of 32,767 characters.'
+  public static ResultTooLong = 'Result exceeds the maximum allowed length.'
   public static FunctionName = (arg: string) => `Function name ${arg} not recognized.`
   public static NamedExpressionName = (arg: string) => `Named expression ${arg} not recognized.`
   public static LicenseKey = (arg: string) => `License key is ${arg}.`